requirements.txt
PyYAML
requests
jinja2
